requirements.txt
Fabric==1.8.0
Mako==1.0.1
MarkupSafe==0.23
Pillow==2.5.3
SQLAlchemy==0.8.3
Tempita==0.5.2
alembic==0.6.6
apns==1.1.2
boto==2.27.0
decorator==3.4.0
ecdsa==0.10
httplib2==0.9
lockfile==0.9.1
paramiko==1.15.2
pbr==0.10.7
psycopg2==2.5.1
pycrypto==2.6.1
python-daemon==1.6
python-dateutil==2.2
python-memcached==1.53
pytz==2014.10
requests==2.1.0
six==1.9.0
sqlalchemy-migrate==0.8.2
tornado==3.1.1
twilio==3.7.2
wsgiref==0.1.2
